Added rudimentary tests for Stat; added walk to the lib
* simplified the Walk/WalkFunc interface to use Info directly
* teach Walk to invoke Readdirnames() and call Lstat() separately
* use Stat/Lstat throughout Walk in lieu of os.Stat/os.Lstat

+++ marshal_test.go
@@ -229,7 +229,7 @@ func randInfo() *Info {
 	}
 
 	if rand.Uint32()&1 > 0 {
-		ix.Mod = uint32(fs.ModeDir)
+		ix.Mod = fs.ModeDir
 	}
 
 	ix.Mod |= 0600
